--- packages/core/src/interfaces/access-token.ts
@@ -1,9 +1,0 @@
-import type { Types } from 'mongoose';
-
-export type IAccessToken = {
-  userId: Types.ObjectId,
-  tokenHash: string,
-  expiredAt: Date,
-  scope: string[],
-  description: string,
-}
